Factorize header URL and height (geor/geor#3134)
Height and URL are read from georchestra datadir

See:
* https://github.com/georchestra/georchestra/issues/3134
* https://github.com/georchestra/geoserver/commit/ab98b34c3c

The implementation is not perfect, I'd have liked to benefit from spring
injection, but I did not wanted to take the risk of loading the
spring-wicket jar in the classpath. There must be a better approach
though, relying on:
https://cwiki.apache.org/confluence/display/WICKET/Spring

Also, I would have expected some places in the GS code where a similar
need is used (accessing spring objects from Wicket) but I could not find
any.

Tests:
Tested in a minimal compo with a header, changing the default.properties
from the datadir, OK