HACK: Avoid slaves to change the updateSequence on reload
Master - slave(s) set up sharing a data directory,
with an external watcher process looking for updateSequence
changes in global.xml and calling slaves' /rest/reload,
leads to a loop of configuration changes as each slave will
update the sequence upon almost every catalog change when
reloading the config.

This patch is a hack to handle that particular situation
and has not been tested on other deployment scenarios.