[GEOS-10074] GeoFence "Admin rules" grant "ADMIN" access to unauthorized users
NOTE: this is a cherry-pick from an unmerged pull request upstream.
Once [1] is merged and integrated upstream to the version georchestra
depends on, this commit is no longer needed.

[1] https://github.com/geoserver/geoserver/pull/5109/commits

-----

- Fix `GeoFenceAccessManager.isWorkspaceAdmin()`: When a user name is
  not provided, using a `RuleFilter` with `SpecialFilterType.ANY`
instead of `SpecialFilterType.DEFAULT` results in admin access to all
workspaces that have an AdminRule set, regardless of the user/role.

- Add `AdminRule` tests, introduce `GeofenceIntegrationTestSupport`
  class to group utility methods and setup/cleanup.

Update GeofenceAccessManager.java